+++ lib/Braintree/TransactionGateway.php
@@ -131,7 +131,12 @@ class TransactionGateway
                 [
                     'eciFlag',
                     'cavv',
-                    'xid'],
+                    'xid',
+                    'threeDSecureVersion',
+                    'authenticationResponse',
+                    'directoryResponse',
+                    'cavvAlgorithm',
+                    'dsTransactionId'],
             ],
             ['options' =>
                 [
@@ -194,6 +199,7 @@ class TransactionGateway
                             'lodgingCheckOutDate',
                             'lodgingName',
                             'roomRate',
+                            'roomTax',
                             'passengerFirstName',
                             'passengerLastName',
                             'passengerMiddleInitial',
@@ -208,6 +214,10 @@ class TransactionGateway
                             'feeAmount',
                             'taxAmount',
                             'restrictedTicket',
+                            'noShow',
+                            'advancedDeposit',
+                            'fireSafe',
+                            'propertyPhone',
                             ['legs' =>
                                 [
                                     'conjunctionTicket',
@@ -227,6 +237,12 @@ class TransactionGateway
                                     'feeAmount',
                                     'taxAmount',
                                     'endorsementOrRestrictions'
+                                ]
+                            ],
+                            ['additionalCharges' =>
+                                [
+                                    'kind',
+                                    'amount'
                                 ]
                             ]
                         ]